requirement.txt
certifi==2021.5.30
chardet==4.0.0
ci-info==0.2.0
click==8.0.1
configobj==5.0.6
configparser==5.0.2
decorator==5.0.9
docx==0.2.4
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
future==0.18.2
httplib2==0.19.1
idna==2.10
importlib-metadata==4.6.0
isodate==0.6.0
lxml==4.3.0
networkx==2.0
nibabel==3.2.1
nipype==1.6.1
numpy==1.21.0
packaging==21.0
pandas==1.3.0
pathlib==1.0.1
Pillow==8.3.0
prov==2.0.0
pydot==1.4.2
PyMuPDF==1.18.14
pyparsing==2.4.7
python-dateutil==2.8.1
python-docx==0.8.11
pytz==2021.1
pyxnat==1.4
rdflib==5.0.0
requests==2.25.1
scipy==1.7.0
simplejson==3.17.2
six==1.16.0
traits==6.2.0
typing-extensions==3.10.0.0
urllib3==1.26.6
xlrd==1.2.0
zipp==3.5.0
